Mbed TLS makefile generator
Create a build directory with subdirectories and symlinks.
Generate a makefile to build the library, sample programs and unit
tests.

Support both in-tree and out-of-tree builds.

Collect dependencies by looking for #include directives without
consideration for the current configuration, so that the resulting
makefile doesn't need to be regenerated when the configuration
changes.

Support changing the configuration when creating the build directory.
This can be done either by copying the configuration file and
modifying it, or by including the original configuratiom and modifying
it through additional #define and #undef.