Make zeroize test independent of the zeroize.c size or structure
The zeroization test is based on a scripted GDB run, which assumes
hard-coded line to break the execution for mbedtls_zeroize operation
evaluation.

The hardcoded break line has already had to be modified more than once
(due to changes made to zeroize.c program), surprising at least one of
the programmers. Therefore, this PR proposes inclusion of an automated
selection of the line for the said break point.